improve loading speed ( async assets loading, chunking node_modules and components, caching in nginx ) ; fix workers usage ; update project to use vue/cli 4 (with yarn) ; add AGPLv3 License

+++ src/services/socket.js
@@ -1,36 +1,31 @@
+/* Copyright 2019 Edouard Maleix, read LICENSE */
+
 import mqtt from 'async-mqtt';
 import logger from './logger';
 import PubSub from './PubSub';
 
 const Storage = window.sessionStorage;
-
 const brokerUrl = process.env.VUE_APP_BROKER_URL;
+const socket = {};
 const baseOptions = {
   keepalive: 60,
   reschedulePings: true,
   protocolId: 'MQTT',
   protocolVersion: 4,
   reconnectPeriod: 1000,
-  connectTimeout: 30 * 1000,
+  connectTimeout: 2 * 1000,
   clean: true,
   clientId: null,
   username: null,
   password: null,
-  // will: { topic: `${clientId}/status`, payload: 'KO?', retain: false, qos: 0 },
 };
 
-const socket = {};
-
 const setSocketId = socketId => {
-  if (Storage) {
-    Storage.setItem('socket-id', socketId);
-  }
+  if (Storage) Storage.setItem('socket-id', socketId);
 };
 
 const delSocketId = () => {
-  if (Storage) {
-    Storage.removeItem('socket-id');
-  }
+  if (Storage) Storage.removeItem('socket-id');
 };
 
 const getSocketId = () => {
@@ -66,42 +61,49 @@ socket.removeToken = () => {
     delete socket.token;
     if (socket.client) {
       PubSub.removeListeners(socket.client);
-      socket.client.end();
+      socket.client.end(true);
+      delete socket.client;
     }
     logger.publish(3, 'socket', 'removeToken:res', 'success');
-    // return;
   } catch (error) {
     logger.publish(3, 'socket', 'removeToken:err', error);
-    // throw error;
   }
+};
+
+const handleMessage = (packet, cb) => {
+  PubSub.handler(packet.topic, packet.payload)
+    .then(() => {
+      cb();
+    })
+    .catch(() => {
+      cb();
+    });
 };
 
 socket.initSocket = async options => {
   try {
     logger.publish(3, 'socket', 'initSocket:req', options);
-
     let socketId = getSocketId();
     if (socketId && socketId !== null && socket.client) {
       return socket;
     }
+
     socket.client = await mqtt.connectAsync(brokerUrl, options);
     logger.publish(3, 'socket', 'onConnect', 'success');
-
-    // socket.client = mqtt.connect(brokerUrl, options);
-    socket.client.on('connect', async state => {
-      logger.publish(3, 'socket', 'onConnect', state);
-      setSocketId(options.clientId);
-    });
-
     setSocketId(options.clientId);
     await PubSub.setListeners(socket.client, socket.token);
 
-    socket.client.on('offline', () => {
-      logger.publish(3, 'socket', 'onDisconnect', '');
-      delSocketId();
+    socket.client.on('reconnect', () => {
+      logger.publish(3, 'socket', 'onReconnecting', options.clientId);
+      // setSocketId(options.clientId);
     });
 
-    socket.client.on('message', PubSub.handler);
+    socket.client.once('offline', () => {
+      logger.publish(3, 'socket', 'onDisconnect', options.clientId);
+      // delSocketId();
+    });
+
+    socket.client.handleMessage = handleMessage;
 
     return socket;
   } catch (error) {
